requirements.txt
fastapi==0.78.0
nebula3-python==3.1.0